+++ src/flamenco-manager/main.go
@@ -22,7 +22,7 @@ import (
 	"github.com/gorilla/mux"
 )
 
-const FLAMENCO_VERSION = "2.0.9"
+const FLAMENCO_VERSION = "2.0.10"
 
 // MongoDB session
 var session *mgo.Session